Consume a FASTQ file and collate basic characteristic observations

This method is a fast C++ implementation of a FASTQ parser aimed to perform a quick collation of summary_statistic information that may be used to make decisions as to how FASTQ sequence collections are crafted.

fishy_fastq(fastq)

Value

DataFrame of FASTQ based sequence observations

Examples

fastq <- system.file("extdata", "example.fastq.gz", package = "floundeR")
fishy_fastq(fastq)

#> (s/q)Lengths==481/8.57659
#> (s/q)Lengths==700/8.93881
#> (s/q)Lengths==722/8.70646
#> (s/q)Lengths==402/8.68473
#> (s/q)Lengths==840/8.76531
#> (s/q)Lengths==351/8.55216
#> (s/q)Lengths==780/9.18107
#> (s/q)Lengths==784/9.0709
#> (s/q)Lengths==803/8.67162
#> (s/q)Lengths==769/8.45845
#> (s/q)Lengths==782/8.56116
#> (s/q)Lengths==780/9.13463
#> (s/q)Lengths==1190/9.15195
#> (s/q)Lengths==1791/8.85981
#> (s/q)Lengths==725/8.57502
#> (s/q)Lengths==726/9.16606
#> (s/q)Lengths==754/9.55239
#> (s/q)Lengths==780/8.88568
#> (s/q)Lengths==752/9.17452
#> (s/q)Lengths==678/8.96442
#> (s/q)Lengths==546/8.7696
#> (s/q)Lengths==307/8.28861
#> (s/q)Lengths==1096/8.79804
#> (s/q)Lengths==728/8.62111
#> (s/q)Lengths==601/9.12606
#> (s/q)Lengths==909/8.61081
#> (s/q)Lengths==1122/9.25246
#> (s/q)Lengths==1000/8.64944
#> (s/q)Lengths==482/9.87446
#> (s/q)Lengths==3106/9.29854
#> (s/q)Lengths==4027/8.63772
#> (s/q)Lengths==1647/8.73716
#> (s/q)Lengths==1667/9.41905
#> (s/q)Lengths==1782/8.70246
#> (s/q)Lengths==1268/8.59724
#> (s/q)Lengths==1287/8.83191
#> (s/q)Lengths==1109/9.55941
#> (s/q)Lengths==516/8.82999
#> (s/q)Lengths==519/9.96519
#> (s/q)Lengths==718/9.2397
#> (s/q)Lengths==1799/8.7873
#> (s/q)Lengths==1543/8.67574
#> (s/q)Lengths==1558/9.04116
#> (s/q)Lengths==970/8.54946
#> (s/q)Lengths==618/9.51802
#> (s/q)Lengths==684/8.38549
#> (s/q)Lengths==872/8.62442
#> (s/q)Lengths==860/8.70863
#> (s/q)Lengths==877/9.15234
#> (s/q)Lengths==761/8.53734
#> (s/q)Lengths==779/8.79721
#> (s/q)Lengths==624/8.76616
#> (s/q)Lengths==640/8.78104
#> (s/q)Lengths==499/8.58515
#> (s/q)Lengths==1634/8.9007
#> (s/q)Lengths==2053/9.11044
#> (s/q)Lengths==561/9.03331
#> (s/q)Lengths==528/9.71094
#> (s/q)Lengths==544/9.13839
#> (s/q)Lengths==495/9.23035
#> (s/q)Lengths==555/9.08794
#> (s/q)Lengths==545/8.66721
#> (s/q)Lengths==352/9.16827
#> (s/q)Lengths==534/8.89984
#> (s/q)Lengths==1022/8.584
#> (s/q)Lengths==1452/8.91447
#> (s/q)Lengths==705/8.78194
#> (s/q)Lengths==629/8.56175
#> (s/q)Lengths==566/8.79156
#> (s/q)Lengths==627/8.99213
#> (s/q)Lengths==630/8.96665
#> (s/q)Lengths==564/8.93914
#> (s/q)Lengths==555/8.90208
#> (s/q)Lengths==558/8.58799
#> (s/q)Lengths==895/8.37915
#> (s/q)Lengths==528/9.55102
#> (s/q)Lengths==583/9.45162
#> (s/q)Lengths==611/9.13133
#> (s/q)Lengths==828/8.57243
#> (s/q)Lengths==1766/9.43075
#> (s/q)Lengths==1237/9.28955
#> (s/q)Lengths==693/8.81142
#> (s/q)Lengths==382/8.89404
#> (s/q)Lengths==963/8.75427
#> (s/q)Lengths==805/8.82375
#> (s/q)Lengths==1163/8.95122
#> (s/q)Lengths==970/8.84356
#> (s/q)Lengths==925/9.1686
#> (s/q)Lengths==949/8.69082
#> (s/q)Lengths==932/8.88989
#> (s/q)Lengths==940/8.83245
#> (s/q)Lengths==901/8.92324
#> (s/q)Lengths==936/8.85932
#> (s/q)Lengths==943/8.87078
#> (s/q)Lengths==980/9.18806
#> (s/q)Lengths==907/8.806
#> (s/q)Lengths==878/9.35706
#> (s/q)Lengths==739/8.72293
#> (s/q)Lengths==1254/8.7334
#> (s/q)Lengths==662/8.75358
#> (s/q)Lengths==2203/9.29464
#> (s/q)Lengths==664/9.00935
#> (s/q)Lengths==675/9.15963
#> (s/q)Lengths==1213/8.86625
#> (s/q)Lengths==1236/9.02684
#> (s/q)Lengths==1068/9.26347
#> (s/q)Lengths==1762/9.44529
#> (s/q)Lengths==1569/9.04412
#> (s/q)Lengths==1021/8.53077
#> (s/q)Lengths==987/9.53164
#> (s/q)Lengths==3278/9.18806
#> (s/q)Lengths==758/8.68972
#> (s/q)Lengths==1758/8.79559
#> (s/q)Lengths==916/8.97886
#> (s/q)Lengths==1015/9.26268
#> (s/q)Lengths==2092/9.27189
#> (s/q)Lengths==2024/9.27361
#> (s/q)Lengths==2034/9.33411
#> (s/q)Lengths==799/8.5147
#> (s/q)Lengths==961/8.80742
#> (s/q)Lengths==532/8.71338
#> (s/q)Lengths==2803/9.11144
#> (s/q)Lengths==691/9.15399
#> (s/q)Lengths==1809/8.61064
#> (s/q)Lengths==1431/9.2437
#> (s/q)Lengths==1730/8.67288
#> (s/q)Lengths==1829/8.63137
#> (s/q)Lengths==1740/8.87871
#> (s/q)Lengths==333/8.6606
#> (s/q)Lengths==618/8.80984
#> (s/q)Lengths==874/8.59846
#> (s/q)Lengths==1149/8.61664
#> (s/q)Lengths==881/9.02753
#> (s/q)Lengths==1792/9.15487
#> (s/q)Lengths==538/8.85055
#> (s/q)Lengths==459/8.98501
#> (s/q)Lengths==588/8.67477
#> (s/q)Lengths==831/8.67305
#> (s/q)Lengths==563/9.23986
#> (s/q)Lengths==606/9.28495
#> (s/q)Lengths==1330/8.67521
#> (s/q)Lengths==475/9.60169
#> (s/q)Lengths==651/9.31855
#> (s/q)Lengths==954/8.61474
#> (s/q)Lengths==562/8.59328
#> (s/q)Lengths==1906/8.95294
#> (s/q)Lengths==616/8.63664
#> (s/q)Lengths==1473/8.83908
#> (s/q)Lengths==868/8.43916
#> (s/q)Lengths==1296/8.96246
#> (s/q)Lengths==586/8.88085
#> (s/q)Lengths==3099/9.04674
#> (s/q)Lengths==1174/8.91595
#> (s/q)Lengths==1040/8.89686
#> (s/q)Lengths==3058/8.76311
#> (s/q)Lengths==956/8.97846
#> (s/q)Lengths==2257/8.72505
#> (s/q)Lengths==2713/9.0669
#> (s/q)Lengths==2922/8.39934
#> (s/q)Lengths==2612/8.67742
#> (s/q)Lengths==2193/8.95901
#> (s/q)Lengths==2705/9.42626
#> (s/q)Lengths==748/8.67452
#> (s/q)Lengths==699/9.22083
#> (s/q)Lengths==1150/8.67922
#> (s/q)Lengths==928/8.74425
#> (s/q)Lengths==931/8.94528
#> (s/q)Lengths==899/9.12102
#> (s/q)Lengths==1002/8.91864
#> (s/q)Lengths==2678/8.91328
#> (s/q)Lengths==912/8.39795
#> (s/q)Lengths==1511/8.86321
#> (s/q)Lengths==2072/8.80654
#> (s/q)Lengths==2197/8.74832
#> (s/q)Lengths==1411/9.36925
#> (s/q)Lengths==1231/8.6524
#> (s/q)Lengths==1101/8.86945
#> (s/q)Lengths==1681/9.04478
#> (s/q)Lengths==1134/9.89522
#> (s/q)Lengths==1726/8.65959
#> (s/q)Lengths==1670/8.9243
#> (s/q)Lengths==1338/8.71656
#> (s/q)Lengths==1692/8.92295
#> (s/q)Lengths==1507/8.88079
#> (s/q)Lengths==1659/8.96982
#> (s/q)Lengths==1740/8.95331
#> (s/q)Lengths==432/8.81659
#> (s/q)Lengths==1741/8.79951
#> (s/q)Lengths==1704/8.6809
#> (s/q)Lengths==1476/8.7198
#> (s/q)Lengths==1734/8.93298
#> (s/q)Lengths==1676/9.14772
#> (s/q)Lengths==1090/8.95649
#> (s/q)Lengths==1083/9.11611
#> (s/q)Lengths==1237/9.16985
#> (s/q)Lengths==758/8.60354
#> (s/q)Lengths==702/9.06111
#> (s/q)Lengths==700/8.61497
#> (s/q)Lengths==592/8.77095
#> (s/q)Lengths==614/8.58288
#> (s/q)Lengths==1108/8.92205
#> (s/q)Lengths==3317/8.69978
#> (s/q)Lengths==767/8.97354
#> (s/q)Lengths==2635/8.90327
#> (s/q)Lengths==1895/8.87791
#> (s/q)Lengths==1945/8.67098
#> (s/q)Lengths==586/8.85442
#> (s/q)Lengths==1175/8.86061
#> (s/q)Lengths==1459/8.83262
#> (s/q)Lengths==1449/8.92139
#> (s/q)Lengths==2060/9.01654
#> (s/q)Lengths==2034/9.4655
#> (s/q)Lengths==1407/9.17365
#> (s/q)Lengths==1545/8.56114
#> (s/q)Lengths==1140/8.89732
#> (s/q)Lengths==805/8.91983
#> (s/q)Lengths==1105/8.95717
#> (s/q)Lengths==1084/9.14708
#> (s/q)Lengths==734/8.44227
#> (s/q)Lengths==704/8.64968
#> (s/q)Lengths==496/8.77574
#> (s/q)Lengths==488/8.8102
#> (s/q)Lengths==360/8.76561
#> (s/q)Lengths==851/9.09471
#> (s/q)Lengths==993/8.67601
#> (s/q)Lengths==1855/9.15971
#> (s/q)Lengths==514/8.59048
#> (s/q)Lengths==1521/8.92366
#> (s/q)Lengths==562/8.52986
#> (s/q)Lengths==1329/9.10441
#> (s/q)Lengths==601/8.94302
#> (s/q)Lengths==1073/8.83993
#> (s/q)Lengths==534/8.83543
#> (s/q)Lengths==365/8.5125
#> (s/q)Lengths==2791/9.13124
#> (s/q)Lengths==512/8.39198
#> (s/q)Lengths==1835/8.73187
#> (s/q)Lengths==1010/9.24041
#> (s/q)Lengths==1089/8.54405
#> (s/q)Lengths==1062/9.01639
#> (s/q)Lengths==1060/8.95262
#> (s/q)Lengths==1113/8.68629
#> (s/q)Lengths==1091/8.86018
#> (s/q)Lengths==1079/8.8805
#> (s/q)Lengths==1050/8.84429
#> (s/q)Lengths==1056/9.33645
#> (s/q)Lengths==1033/8.84716
#> (s/q)Lengths==483/8.93382
#> (s/q)Lengths==1796/8.56609
#> (s/q)Lengths==1816/9.04046
#> (s/q)Lengths==911/8.81741
#> (s/q)Lengths==651/8.93711
#> (s/q)Lengths==1295/8.9591
#> (s/q)Lengths==1521/8.74907
#> (s/q)Lengths==1557/8.90261
#> (s/q)Lengths==1118/9.06672
#> (s/q)Lengths==1218/8.65449
#> (s/q)Lengths==1210/9.03194
#> (s/q)Lengths==923/8.52245
#> (s/q)Lengths==1059/8.93238
#> (s/q)Lengths==944/8.56471
#> (s/q)Lengths==1597/9.12289
#> (s/q)Lengths==1357/8.9702
#> (s/q)Lengths==1289/9.09617
#> (s/q)Lengths==1303/9.23113
#> (s/q)Lengths==846/9.00042
#> (s/q)Lengths==1044/8.50588
#> (s/q)Lengths==1258/8.58069
#> (s/q)Lengths==563/9.11458
#> (s/q)Lengths==844/8.66403
#> (s/q)Lengths==1606/8.61322
#> (s/q)Lengths==1037/8.60679
#> (s/q)Lengths==1378/8.74075
#> (s/q)Lengths==1923/9.18391
#> (s/q)Lengths==1551/8.76812
#> (s/q)Lengths==1069/8.78529
#> (s/q)Lengths==965/9.15791
#> (s/q)Lengths==516/8.59999
#> (s/q)Lengths==1354/8.57554
#> (s/q)Lengths==612/8.59438
#> (s/q)Lengths==342/9.19475
#> (s/q)Lengths==1035/9.01736
#> (s/q)Lengths==1154/9.0386
#> (s/q)Lengths==1365/9.03866
#> (s/q)Lengths==2311/8.83905
#> (s/q)Lengths==858/9.22975
#> (s/q)Lengths==824/8.87256
#> (s/q)Lengths==561/9.47291
#> (s/q)Lengths==2034/8.89998
#> (s/q)Lengths==389/9.09217
#> (s/q)Lengths==1305/9.01045
#> (s/q)Lengths==1237/9.15461
#> (s/q)Lengths==531/8.82519
#> (s/q)Lengths==486/8.94583
#> (s/q)Lengths==2824/8.67645
#> (s/q)Lengths==1276/8.98047
#> (s/q)Lengths==1258/8.9954
#> (s/q)Lengths==966/8.96331
#> (s/q)Lengths==871/8.95946
#> (s/q)Lengths==855/9.05873
#> (s/q)Lengths==871/8.96391
#> (s/q)Lengths==836/9.27578
#> (s/q)Lengths==1284/8.83685
#> (s/q)Lengths==896/8.5015
#> (s/q)Lengths==1654/8.74294
#> (s/q)Lengths==729/9.04378
#> (s/q)Lengths==335/8.67484
#> (s/q)Lengths==922/8.93263
#> (s/q)Lengths==823/8.96256
#> (s/q)Lengths==753/9.26775
#> (s/q)Lengths==893/8.6873
#> (s/q)Lengths==1120/8.71758
#> (s/q)Lengths==1151/9.05668
#> (s/q)Lengths==640/8.98785
#> (s/q)Lengths==1105/8.7628
#> (s/q)Lengths==1142/9.61086
#> (s/q)Lengths==1137/8.80171
#> (s/q)Lengths==1163/8.68627
#> (s/q)Lengths==1150/8.62351
#> (s/q)Lengths==1164/8.87418
#> (s/q)Lengths==1141/8.78061
#> (s/q)Lengths==1148/9.01564
#> (s/q)Lengths==1139/8.77163
#> (s/q)Lengths==1139/8.56305
#> (s/q)Lengths==1131/8.97196
#> (s/q)Lengths==943/8.92553
#> (s/q)Lengths==1151/9.03409
#> (s/q)Lengths==1109/9.15602
#> (s/q)Lengths==1129/9.02799
#> (s/q)Lengths==1151/8.94778
#> (s/q)Lengths==1130/8.59872
#> (s/q)Lengths==1124/9.16041
#> (s/q)Lengths==1128/8.98414
#> (s/q)Lengths==1137/8.71184
#> (s/q)Lengths==921/8.94284
#> (s/q)Lengths==633/9.00619
#> (s/q)Lengths==1115/8.47824
#> (s/q)Lengths==1144/8.59329
#> (s/q)Lengths==1132/9.03281
#> (s/q)Lengths==1132/8.74908
#> (s/q)Lengths==1115/8.91222
#> (s/q)Lengths==1128/8.88596
#> (s/q)Lengths==1147/9.0841
#> (s/q)Lengths==1134/9.27695
#> (s/q)Lengths==1102/8.63063
#> (s/q)Lengths==1136/9.56229
#> (s/q)Lengths==1116/8.77578
#> (s/q)Lengths==1090/8.80201
#> (s/q)Lengths==1117/8.72804
#> (s/q)Lengths==615/8.86621
#> (s/q)Lengths==1936/8.91841
#> (s/q)Lengths==477/9.30661
#> (s/q)Lengths==479/8.5854
#> (s/q)Lengths==462/8.61817
#> (s/q)Lengths==2538/8.77
#> (s/q)Lengths==5946/9.23209
#> (s/q)Lengths==2992/9.06129
#> (s/q)Lengths==2453/9.60696
#> (s/q)Lengths==3141/8.7643
#> (s/q)Lengths==3159/9.01359
#> (s/q)Lengths==1967/8.97291
#> (s/q)Lengths==2136/8.63392
#> (s/q)Lengths==378/8.90428
#> (s/q)Lengths==1036/9.3953
#> (s/q)Lengths==949/8.96007
#> (s/q)Lengths==1006/8.8449
#> (s/q)Lengths==1008/8.71123
#> (s/q)Lengths==985/8.91994
#> (s/q)Lengths==978/8.56613
#> (s/q)Lengths==975/8.84439
#> (s/q)Lengths==969/9.21099
#> (s/q)Lengths==1003/8.75638
#> (s/q)Lengths==863/8.79231
#> (s/q)Lengths==684/8.9815
#> (s/q)Lengths==677/9.0918
#> (s/q)Lengths==658/8.7351
#> (s/q)Lengths==984/8.80264
#> (s/q)Lengths==672/8.84894
#> (s/q)Lengths==681/8.85822
#> (s/q)Lengths==653/9.40076
#> (s/q)Lengths==659/8.89543
#> (s/q)Lengths==652/8.92652
#> (s/q)Lengths==657/8.76244
#> (s/q)Lengths==674/9.05243
#> (s/q)Lengths==669/8.72265
#> (s/q)Lengths==675/8.74356
#> (s/q)Lengths==660/9.07441
#> (s/q)Lengths==642/8.49316
#> (s/q)Lengths==631/8.70215
#> (s/q)Lengths==691/9.08494
#> (s/q)Lengths==663/8.86092
#> (s/q)Lengths==679/9.06759
#> (s/q)Lengths==652/8.97812
#> (s/q)Lengths==656/8.95125
#> (s/q)Lengths==660/8.98206
#> (s/q)Lengths==682/9.24732
#> (s/q)Lengths==681/8.90795
#> (s/q)Lengths==697/8.85469
#> (s/q)Lengths==655/9.15375
#> (s/q)Lengths==609/8.86356
#> (s/q)Lengths==639/9.22009
#> (s/q)Lengths==659/8.53222
#> (s/q)Lengths==669/8.95477
#> (s/q)Lengths==663/8.51285
#> (s/q)Lengths==646/9.34239
#> (s/q)Lengths==691/8.88967
#> (s/q)Lengths==662/9.12567
#> (s/q)Lengths==687/9.00701
#> (s/q)Lengths==690/9.17183
#> (s/q)Lengths==583/8.56301
#> (s/q)Lengths==534/9.37602
#> (s/q)Lengths==2681/8.86669
#> (s/q)Lengths==429/8.79625
#> (s/q)Lengths==1115/8.83632
#> (s/q)Lengths==1104/8.89966
#> (s/q)Lengths==1379/8.78006
#> (s/q)Lengths==1752/8.71021
#> (s/q)Lengths==1090/9.21502
#> (s/q)Lengths==1564/8.68507
#> (s/q)Lengths==1578/8.96585
#> (s/q)Lengths==1280/8.91588
#> (s/q)Lengths==2345/8.80521
#> (s/q)Lengths==1571/8.71109
#> (s/q)Lengths==1514/8.86068
#> (s/q)Lengths==1525/8.79345
#> (s/q)Lengths==1333/9.26931
#> (s/q)Lengths==2111/9.17134
#> (s/q)Lengths==1900/8.91572
#> (s/q)Lengths==767/8.37928
#> (s/q)Lengths==1741/8.74423
#> (s/q)Lengths==1277/8.91618
#> (s/q)Lengths==1010/8.93635
#> (s/q)Lengths==943/8.69858
#> (s/q)Lengths==1904/9.21287
#> (s/q)Lengths==2399/8.80808
#> (s/q)Lengths==2392/9.00692
#> (s/q)Lengths==549/8.85714
#> (s/q)Lengths==499/9.22622
#> (s/q)Lengths==520/8.89154
#> (s/q)Lengths==637/9.48892
#> (s/q)Lengths==526/9.09899
#> (s/q)Lengths==476/9.64249
#> (s/q)Lengths==553/8.89378
#> (s/q)Lengths==539/9.40377
#> (s/q)Lengths==571/8.78327
#> (s/q)Lengths==1317/8.96794
#> (s/q)Lengths==530/9.08378
#> (s/q)Lengths==550/9.01677
#> (s/q)Lengths==549/9.13356
#> (s/q)Lengths==695/9.58309
#> (s/q)Lengths==353/8.42354
#> (s/q)Lengths==948/9.22515
#> (s/q)Lengths==906/8.72944
#> (s/q)Lengths==523/9.52144
#> (s/q)Lengths==563/8.59668
#> (s/q)Lengths==535/9.05286
#> (s/q)Lengths==1320/8.67511
#> (s/q)Lengths==575/8.20925
#> (s/q)Lengths==562/8.59545
#> (s/q)Lengths==585/8.72153
#> (s/q)Lengths==566/9.05627
#> (s/q)Lengths==565/8.46285
#> (s/q)Lengths==573/8.31997
#> (s/q)Lengths==564/8.81165
#> (s/q)Lengths==577/8.81138
#> (s/q)Lengths==587/8.763
#> (s/q)Lengths==455/8.59235
#> (s/q)Lengths==860/9.3945
#> (s/q)Lengths==1382/8.92901
#> (s/q)Lengths==1358/9.42125
#> (s/q)Lengths==945/9.27784
#> (s/q)Lengths==838/8.8683
#> (s/q)Lengths==1361/8.59663
#> (s/q)Lengths==1995/8.95628
#> (s/q)Lengths==957/8.86721
#> (s/q)Lengths==978/8.9626
#> (s/q)Lengths==954/9.15223
#> (s/q)Lengths==719/8.97553
#> (s/q)Lengths==599/9.12747
#> (s/q)Lengths==357/8.45631
#> (s/q)Lengths==650/8.98059
#> (s/q)Lengths==643/9.42634
#> (s/q)Lengths==571/8.8984
#> (s/q)Lengths==605/8.59362
#> (s/q)Lengths==600/9.17783
#> (s/q)Lengths==570/9.35531
#> (s/q)Lengths==546/8.68603
#> (s/q)Lengths==589/8.80951
#> (s/q)Lengths==607/8.98092
#> (s/q)Lengths==465/9.47996
#> (s/q)Lengths==605/8.61614
#> (s/q)Lengths==604/8.69993
#> (s/q)Lengths==592/8.44073
#> (s/q)Lengths==566/8.78935
#> (s/q)Lengths==1034/9.16976
#> (s/q)Lengths==1057/8.90821
#> (s/q)Lengths==1041/8.87903
#> (s/q)Lengths==1033/9.69885
#> (s/q)Lengths==1036/9.1311
#> (s/q)Lengths==1044/9.77645
#> (s/q)Lengths==1089/8.80954
#> (s/q)Lengths==1025/8.70883
#> (s/q)Lengths==603/8.30092
#> (s/q)Lengths==1018/8.80424
#> (s/q)Lengths==1030/9.07898
#> (s/q)Lengths==1050/8.62511
#> (s/q)Lengths==1032/8.64414
#> (s/q)Lengths==1063/9.02777
#> (s/q)Lengths==173/8.72015
#> (s/q)Lengths==623/8.91577
#> (s/q)Lengths==699/8.46396
#> (s/q)Lengths==1060/9.38204
#> (s/q)Lengths==1006/9.08183
#> (s/q)Lengths==1043/9.12588
#> (s/q)Lengths==695/8.83201
#> (s/q)Lengths==978/8.78203
#> (s/q)Lengths==825/8.61356
#> (s/q)Lengths==804/8.68367
#> (s/q)Lengths==1578/8.51855
#> (s/q)Lengths==1254/8.70991
#> (s/q)Lengths==1241/8.90236
#> (s/q)Lengths==1233/8.97173
#> (s/q)Lengths==532/9.1599
#> (s/q)Lengths==764/9.56652
#> (s/q)Lengths==764/8.56798
#> (s/q)Lengths==775/8.92736
#> (s/q)Lengths==1373/8.67609
#> (s/q)Lengths==856/8.88209
#> (s/q)Lengths==770/9.18359
#> (s/q)Lengths==1240/8.96098
#> (s/q)Lengths==789/9.05123
#> (s/q)Lengths==413/9.33406
#> (s/q)Lengths==2629/9.28716
#> (s/q)Lengths==315/8.44129
#> (s/q)Lengths==330/8.99583
#> (s/q)Lengths==330/8.61427
#> (s/q)Lengths==312/8.19371
#> (s/q)Lengths==441/9.17186
#> (s/q)Lengths==3215/9.01259
#> (s/q)Lengths==1216/9.05857
#> (s/q)Lengths==1845/8.95943
#> (s/q)Lengths==560/8.76089
#> (s/q)Lengths==1017/8.82238
#> (s/q)Lengths==1517/8.82508
#> (s/q)Lengths==795/8.57381
#> (s/q)Lengths==800/8.83117
#> (s/q)Lengths==272/8.44073
#> (s/q)Lengths==302/8.40364
#> (s/q)Lengths==313/8.45987
#> (s/q)Lengths==2118/8.80146
#> (s/q)Lengths==307/9.15053
#> (s/q)Lengths==290/9.08429
#> (s/q)Lengths==1156/8.88622
#> (s/q)Lengths==1056/8.73305
#> (s/q)Lengths==1064/9.3329
#> (s/q)Lengths==2150/8.96734
#> (s/q)Lengths==708/8.70974
#> (s/q)Lengths==1076/8.83439
#> (s/q)Lengths==2047/8.58326
#> (s/q)Lengths==1952/8.86574
#> (s/q)Lengths==791/9.07079
#> (s/q)Lengths==1620/9.11558
#> (s/q)Lengths==1636/8.82113
#> (s/q)Lengths==1297/8.97041
#> (s/q)Lengths==1273/9.03323
#> (s/q)Lengths==1262/8.8735
#> (s/q)Lengths==1302/8.77994
#> (s/q)Lengths==1289/8.66411
#> (s/q)Lengths==1258/8.40794
#> (s/q)Lengths==1285/8.84717
#> (s/q)Lengths==1259/8.69594
#> (s/q)Lengths==1499/8.82788
#> (s/q)Lengths==1475/8.66154
#> (s/q)Lengths==596/9.56807
#> (s/q)Lengths==548/8.57869
#> (s/q)Lengths==532/8.55538
#> (s/q)Lengths==598/9.43539
#> (s/q)Lengths==581/8.70029
#> (s/q)Lengths==526/8.7238
#> (s/q)Lengths==550/8.67231
#> (s/q)Lengths==538/9.05694
#> (s/q)Lengths==574/8.85959
#> (s/q)Lengths==872/9.37351
#> (s/q)Lengths==613/8.38877
#> (s/q)Lengths==335/8.5798
#> (s/q)Lengths==449/9.22073
#> (s/q)Lengths==783/8.93523
#> (s/q)Lengths==647/9.20783
#> (s/q)Lengths==986/9.22083
#> (s/q)Lengths==1711/9.45774
#> (s/q)Lengths==1822/8.83712
#> (s/q)Lengths==648/8.81702
#> (s/q)Lengths==809/8.76186
#> (s/q)Lengths==1346/9.23652
#> (s/q)Lengths==642/8.88918
#> (s/q)Lengths==517/8.70805
#> (s/q)Lengths==1080/8.61504
#> (s/q)Lengths==628/9.02887
#> (s/q)Lengths==662/8.72771
#> (s/q)Lengths==1119/8.91417
#> (s/q)Lengths==1168/9.18803
#> (s/q)Lengths==425/8.5196
#> (s/q)Lengths==385/8.78708
#> (s/q)Lengths==849/8.66891
#> (s/q)Lengths==1062/8.82809
#> (s/q)Lengths==1266/9.12578
#> (s/q)Lengths==1353/9.35482
#> (s/q)Lengths==653/9.23631
#> (s/q)Lengths==625/8.77072
#> (s/q)Lengths==560/9.12298
#> (s/q)Lengths==1008/8.64416
#> (s/q)Lengths==1194/8.63098
#> (s/q)Lengths==3685/8.99573
#> (s/q)Lengths==1168/8.79359
#> (s/q)Lengths==1217/8.27504
#> (s/q)Lengths==1002/9.13898
#> (s/q)Lengths==1188/8.65541
#> (s/q)Lengths==577/8.79659
#> (s/q)Lengths==555/8.87167
#> (s/q)Lengths==555/8.544
#> (s/q)Lengths==870/8.8305
#> (s/q)Lengths==1101/8.63767
#> (s/q)Lengths==1977/9.29351
#> (s/q)Lengths==2751/9.04612
#> (s/q)Lengths==381/9.24289
#> Reading beyond end of file ...

#>      id length   mean_q
#> 1     0    481 8.576590
#> 2     1    700 8.938814
#> 3     2    722 8.706461
#> 4     3    402 8.684725
#> 5     4    840 8.765314
#> 6     5    351 8.552162
#> 7     6    780 9.181071
#> 8     7    784 9.070897
#> 9     8    803 8.671617
#> 10    9    769 8.458450
#> 11   10    782 8.561163
#> 12   11    780 9.134631
#> 13   12   1190 9.151952
#> 14   13   1791 8.859815
#> 15   14    725 8.575019
#> 16   15    726 9.166060
#> 17   16    754 9.552390
#> 18   17    780 8.885683
#> 19   18    752 9.174519
#> 20   19    678 8.964424
#> 21   20    546 8.769598
#> 22   21    307 8.288609
#> 23   22   1096 8.798040
#> 24   23    728 8.621111
#> 25   24    601 9.126060
#> 26   25    909 8.610812
#> 27   26   1122 9.252459
#> 28   27   1000 8.649443
#> 29   28    482 9.874456
#> 30   29   3106 9.298542
#> 31   30   4027 8.637724
#> 32   31   1647 8.737156
#> 33   32   1667 9.419045
#> 34   33   1782 8.702462
#> 35   34   1268 8.597244
#> 36   35   1287 8.831912
#> 37   36   1109 9.559415
#> 38   37    516 8.829993
#> 39   38    519 9.965189
#> 40   39    718 9.239699
#> 41   40   1799 8.787300
#> 42   41   1543 8.675742
#> 43   42   1558 9.041159
#> 44   43    970 8.549458
#> 45   44    618 9.518017
#> 46   45    684 8.385486
#> 47   46    872 8.624417
#> 48   47    860 8.708627
#> 49   48    877 9.152344
#> 50   49    761 8.537343
#> 51   50    779 8.797213
#> 52   51    624 8.766164
#> 53   52    640 8.781042
#> 54   53    499 8.585148
#> 55   54   1634 8.900698
#> 56   55   2053 9.110444
#> 57   56    561 9.033309
#> 58   57    528 9.710940
#> 59   58    544 9.138390
#> 60   59    495 9.230347
#> 61   60    555 9.087939
#> 62   61    545 8.667213
#> 63   62    352 9.168268
#> 64   63    534 8.899837
#> 65   64   1022 8.583999
#> 66   65   1452 8.914466
#> 67   66    705 8.781941
#> 68   67    629 8.561754
#> 69   68    566 8.791556
#> 70   69    627 8.992131
#> 71   70    630 8.966647
#> 72   71    564 8.939142
#> 73   72    555 8.902081
#> 74   73    558 8.587989
#> 75   74    895 8.379152
#> 76   75    528 9.551015
#> 77   76    583 9.451621
#> 78   77    611 9.131329
#> 79   78    828 8.572431
#> 80   79   1766 9.430747
#> 81   80   1237 9.289551
#> 82   81    693 8.811421
#> 83   82    382 8.894040
#> 84   83    963 8.754269
#> 85   84    805 8.823750
#> 86   85   1163 8.951216
#> 87   86    970 8.843556
#> 88   87    925 9.168600
#> 89   88    949 8.690817
#> 90   89    932 8.889891
#> 91   90    940 8.832452
#> 92   91    901 8.923236
#> 93   92    936 8.859321
#> 94   93    943 8.870780
#> 95   94    980 9.188061
#> 96   95    907 8.805997
#> 97   96    878 9.357055
#> 98   97    739 8.722930
#> 99   98   1254 8.733398
#> 100  99    662 8.753577
#> 101 100   2203 9.294639
#> 102 101    664 9.009345
#> 103 102    675 9.159626
#> 104 103   1213 8.866251
#> 105 104   1236 9.026835
#> 106 105   1068 9.263474
#> 107 106   1762 9.445290
#> 108 107   1569 9.044123
#> 109 108   1021 8.530767
#> 110 109    987 9.531640
#> 111 110   3278 9.188058
#> 112 111    758 8.689724
#> 113 112   1758 8.795593
#> 114 113    916 8.978864
#> 115 114   1015 9.262681
#> 116 115   2092 9.271888
#> 117 116   2024 9.273611
#> 118 117   2034 9.334114
#> 119 118    799 8.514698
#> 120 119    961 8.807415
#> 121 120    532 8.713382
#> 122 121   2803 9.111441
#> 123 122    691 9.153989
#> 124 123   1809 8.610636
#> 125 124   1431 9.243698
#> 126 125   1730 8.672883
#> 127 126   1829 8.631370
#> 128 127   1740 8.878706
#> 129 128    333 8.660597
#> 130 129    618 8.809836
#> 131 130    874 8.598460
#> 132 131   1149 8.616637
#> 133 132    881 9.027526
#> 134 133   1792 9.154873
#> 135 134    538 8.850554
#> 136 135    459 8.985008
#> 137 136    588 8.674768
#> 138 137    831 8.673054
#> 139 138    563 9.239863
#> 140 139    606 9.284954
#> 141 140   1330 8.675208
#> 142 141    475 9.601694
#> 143 142    651 9.318546
#> 144 143    954 8.614737
#> 145 144    562 8.593277
#> 146 145   1906 8.952943
#> 147 146    616 8.636640
#> 148 147   1473 8.839078
#> 149 148    868 8.439160
#> 150 149   1296 8.962461
#> 151 150    586 8.880845
#> 152 151   3099 9.046738
#> 153 152   1174 8.915950
#> 154 153   1040 8.896862
#> 155 154   3058 8.763107
#> 156 155    956 8.978460
#> 157 156   2257 8.725055
#> 158 157   2713 9.066904
#> 159 158   2922 8.399341
#> 160 159   2612 8.677419
#> 161 160   2193 8.959013
#> 162 161   2705 9.426265
#> 163 162    748 8.674518
#> 164 163    699 9.220833
#> 165 164   1150 8.679215
#> 166 165    928 8.744246
#> 167 166    931 8.945282
#> 168 167    899 9.121024
#> 169 168   1002 8.918637
#> 170 169   2678 8.913277
#> 171 170    912 8.397955
#> 172 171   1511 8.863208
#> 173 172   2072 8.806537
#> 174 173   2197 8.748316
#> 175 174   1411 9.369255
#> 176 175   1231 8.652401
#> 177 176   1101 8.869453
#> 178 177   1681 9.044779
#> 179 178   1134 9.895219
#> 180 179   1726 8.659587
#> 181 180   1670 8.924303
#> 182 181   1338 8.716558
#> 183 182   1692 8.922948
#> 184 183   1507 8.880793
#> 185 184   1659 8.969820
#> 186 185   1740 8.953312
#> 187 186    432 8.816593
#> 188 187   1741 8.799508
#> 189 188   1704 8.680899
#> 190 189   1476 8.719795
#> 191 190   1734 8.932984
#> 192 191   1676 9.147722
#> 193 192   1090 8.956491
#> 194 193   1083 9.116109
#> 195 194   1237 9.169847
#> 196 195    758 8.603536
#> 197 196    702 9.061111
#> 198 197    700 8.614972
#> 199 198    592 8.770947
#> 200 199    614 8.582882
#> 201 200   1108 8.922046
#> 202 201   3317 8.699775
#> 203 202    767 8.973541
#> 204 203   2635 8.903270
#> 205 204   1895 8.877907
#> 206 205   1945 8.670979
#> 207 206    586 8.854418
#> 208 207   1175 8.860614
#> 209 208   1459 8.832618
#> 210 209   1449 8.921386
#> 211 210   2060 9.016544
#> 212 211   2034 9.465502
#> 213 212   1407 9.173653
#> 214 213   1545 8.561141
#> 215 214   1140 8.897316
#> 216 215    805 8.919828
#> 217 216   1105 8.957165
#> 218 217   1084 9.147075
#> 219 218    734 8.442271
#> 220 219    704 8.649677
#> 221 220    496 8.775745
#> 222 221    488 8.810199
#> 223 222    360 8.765608
#> 224 223    851 9.094712
#> 225 224    993 8.676013
#> 226 225   1855 9.159714
#> 227 226    514 8.590481
#> 228 227   1521 8.923656
#> 229 228    562 8.529856
#> 230 229   1329 9.104410
#> 231 230    601 8.943021
#> 232 231   1073 8.839932
#> 233 232    534 8.835434
#> 234 233    365 8.512503
#> 235 234   2791 9.131243
#> 236 235    512 8.391978
#> 237 236   1835 8.731873
#> 238 237   1010 9.240406
#> 239 238   1089 8.544054
#> 240 239   1062 9.016390
#> 241 240   1060 8.952624
#> 242 241   1113 8.686292
#> 243 242   1091 8.860182
#> 244 243   1079 8.880504
#> 245 244   1050 8.844290
#> 246 245   1056 9.336447
#> 247 246   1033 8.847164
#> 248 247    483 8.933823
#> 249 248   1796 8.566093
#> 250 249   1816 9.040464
#> 251 250    911 8.817411
#> 252 251    651 8.937114
#> 253 252   1295 8.959100
#> 254 253   1521 8.749072
#> 255 254   1557 8.902612
#> 256 255   1118 9.066720
#> 257 256   1218 8.654493
#> 258 257   1210 9.031938
#> 259 258    923 8.522453
#> 260 259   1059 8.932382
#> 261 260    944 8.564706
#> 262 261   1597 9.122895
#> 263 262   1357 8.970204
#> 264 263   1289 9.096166
#> 265 264   1303 9.231134
#> 266 265    846 9.000420
#> 267 266   1044 8.505878
#> 268 267   1258 8.580690
#> 269 268    563 9.114577
#> 270 269    844 8.664027
#> 271 270   1606 8.613220
#> 272 271   1037 8.606793
#> 273 272   1378 8.740752
#> 274 273   1923 9.183908
#> 275 274   1551 8.768124
#> 276 275   1069 8.785294
#> 277 276    965 9.157911
#> 278 277    516 8.599988
#> 279 278   1354 8.575545
#> 280 279    612 8.594377
#> 281 280    342 9.194752
#> 282 281   1035 9.017363
#> 283 282   1154 9.038597
#> 284 283   1365 9.038660
#> 285 284   2311 8.839053
#> 286 285    858 9.229750
#> 287 286    824 8.872558
#> 288 287    561 9.472907
#> 289 288   2034 8.899984
#> 290 289    389 9.092170
#> 291 290   1305 9.010448
#> 292 291   1237 9.154613
#> 293 292    531 8.825195
#> 294 293    486 8.945825
#> 295 294   2824 8.676455
#> 296 295   1276 8.980473
#> 297 296   1258 8.995400
#> 298 297    966 8.963307
#> 299 298    871 8.959463
#> 300 299    855 9.058735
#> 301 300    871 8.963910
#> 302 301    836 9.275776
#> 303 302   1284 8.836848
#> 304 303    896 8.501498
#> 305 304   1654 8.742943
#> 306 305    729 9.043782
#> 307 306    335 8.674843
#> 308 307    922 8.932634
#> 309 308    823 8.962562
#> 310 309    753 9.267746
#> 311 310    893 8.687301
#> 312 311   1120 8.717583
#> 313 312   1151 9.056684
#> 314 313    640 8.987849
#> 315 314   1105 8.762798
#> 316 315   1142 9.610859
#> 317 316   1137 8.801708
#> 318 317   1163 8.686275
#> 319 318   1150 8.623515
#> 320 319   1164 8.874183
#> 321 320   1141 8.780610
#> 322 321   1148 9.015638
#> 323 322   1139 8.771626
#> 324 323   1139 8.563053
#> 325 324   1131 8.971964
#> 326 325    943 8.925531
#> 327 326   1151 9.034090
#> 328 327   1109 9.156023
#> 329 328   1129 9.027990
#> 330 329   1151 8.947779
#> 331 330   1130 8.598717
#> 332 331   1124 9.160414
#> 333 332   1128 8.984138
#> 334 333   1137 8.711840
#> 335 334    921 8.942837
#> 336 335    633 9.006188
#> 337 336   1115 8.478238
#> 338 337   1144 8.593292
#> 339 338   1132 9.032812
#> 340 339   1132 8.749082
#> 341 340   1115 8.912217
#> 342 341   1128 8.885962
#> 343 342   1147 9.084100
#> 344 343   1134 9.276955
#> 345 344   1102 8.630629
#> 346 345   1136 9.562294
#> 347 346   1116 8.775780
#> 348 347   1090 8.802011
#> 349 348   1117 8.728036
#> 350 349    615 8.866208
#> 351 350   1936 8.918410
#> 352 351    477 9.306606
#> 353 352    479 8.585401
#> 354 353    462 8.618173
#> 355 354   2538 8.770000
#> 356 355   5946 9.232093
#> 357 356   2992 9.061294
#> 358 357   2453 9.606964
#> 359 358   3141 8.764297
#> 360 359   3159 9.013587
#> 361 360   1967 8.972908
#> 362 361   2136 8.633924
#> 363 362    378 8.904277
#> 364 363   1036 9.395297
#> 365 364    949 8.960071
#> 366 365   1006 8.844896
#> 367 366   1008 8.711229
#> 368 367    985 8.919939
#> 369 368    978 8.566126
#> 370 369    975 8.844392
#> 371 370    969 9.210991
#> 372 371   1003 8.756380
#> 373 372    863 8.792306
#> 374 373    684 8.981501
#> 375 374    677 9.091798
#> 376 375    658 8.735104
#> 377 376    984 8.802644
#> 378 377    672 8.848940
#> 379 378    681 8.858216
#> 380 379    653 9.400763
#> 381 380    659 8.895428
#> 382 381    652 8.926518
#> 383 382    657 8.762445
#> 384 383    674 9.052426
#> 385 384    669 8.722652
#> 386 385    675 8.743557
#> 387 386    660 9.074414
#> 388 387    642 8.493161
#> 389 388    631 8.702148
#> 390 389    691 9.084943
#> 391 390    663 8.860919
#> 392 391    679 9.067593
#> 393 392    652 8.978120
#> 394 393    656 8.951246
#> 395 394    660 8.982062
#> 396 395    682 9.247323
#> 397 396    681 8.907953
#> 398 397    697 8.854692
#> 399 398    655 9.153746
#> 400 399    609 8.863561
#> 401 400    639 9.220086
#> 402 401    659 8.532218
#> 403 402    669 8.954770
#> 404 403    663 8.512855
#> 405 404    646 9.342391
#> 406 405    691 8.889667
#> 407 406    662 9.125675
#> 408 407    687 9.007006
#> 409 408    690 9.171832
#> 410 409    583 8.563015
#> 411 410    534 9.376023
#> 412 411   2681 8.866687
#> 413 412    429 8.796249
#> 414 413   1115 8.836318
#> 415 414   1104 8.899658
#> 416 415   1379 8.780057
#> 417 416   1752 8.710213
#> 418 417   1090 9.215020
#> 419 418   1564 8.685073
#> 420 419   1578 8.965851
#> 421 420   1280 8.915881
#> 422 421   2345 8.805212
#> 423 422   1571 8.711089
#> 424 423   1514 8.860684
#> 425 424   1525 8.793452
#> 426 425   1333 9.269311
#> 427 426   2111 9.171345
#> 428 427   1900 8.915723
#> 429 428    767 8.379281
#> 430 429   1741 8.744231
#> 431 430   1277 8.916176
#> 432 431   1010 8.936346
#> 433 432    943 8.698577
#> 434 433   1904 9.212870
#> 435 434   2399 8.808081
#> 436 435   2392 9.006919
#> 437 436    549 8.857136
#> 438 437    499 9.226223
#> 439 438    520 8.891541
#> 440 439    637 9.488916
#> 441 440    526 9.098991
#> 442 441    476 9.642491
#> 443 442    553 8.893782
#> 444 443    539 9.403773
#> 445 444    571 8.783266
#> 446 445   1317 8.967943
#> 447 446    530 9.083784
#> 448 447    550 9.016772
#> 449 448    549 9.133563
#> 450 449    695 9.583085
#> 451 450    353 8.423545
#> 452 451    948 9.225145
#> 453 452    906 8.729439
#> 454 453    523 9.521440
#> 455 454    563 8.596677
#> 456 455    535 9.052860
#> 457 456   1320 8.675113
#> 458 457    575 8.209250
#> 459 458    562 8.595451
#> 460 459    585 8.721528
#> 461 460    566 9.056266
#> 462 461    565 8.462854
#> 463 462    573 8.319966
#> 464 463    564 8.811651
#> 465 464    577 8.811384
#> 466 465    587 8.763002
#> 467 466    455 8.592346
#> 468 467    860 9.394502
#> 469 468   1382 8.929006
#> 470 469   1358 9.421250
#> 471 470    945 9.277841
#> 472 471    838 8.868300
#> 473 472   1361 8.596629
#> 474 473   1995 8.956285
#> 475 474    957 8.867213
#> 476 475    978 8.962595
#> 477 476    954 9.152226
#> 478 477    719 8.975530
#> 479 478    599 9.127474
#> 480 479    357 8.456306
#> 481 480    650 8.980589
#> 482 481    643 9.426341
#> 483 482    571 8.898402
#> 484 483    605 8.593619
#> 485 484    600 9.177834
#> 486 485    570 9.355305
#> 487 486    546 8.686034
#> 488 487    589 8.809509
#> 489 488    607 8.980917
#> 490 489    465 9.479959
#> 491 490    605 8.616143
#> 492 491    604 8.699931
#> 493 492    592 8.440726
#> 494 493    566 8.789350
#> 495 494   1034 9.169764
#> 496 495   1057 8.908205
#> 497 496   1041 8.879029
#> 498 497   1033 9.698854
#> 499 498   1036 9.131104
#> 500 499   1044 9.776451
#> 501 500   1089 8.809536
#> 502 501   1025 8.708831
#> 503 502    603 8.300916
#> 504 503   1018 8.804236
#> 505 504   1030 9.078977
#> 506 505   1050 8.625113
#> 507 506   1032 8.644139
#> 508 507   1063 9.027771
#> 509 508    173 8.720148
#> 510 509    623 8.915773
#> 511 510    699 8.463956
#> 512 511   1060 9.382037
#> 513 512   1006 9.081835
#> 514 513   1043 9.125876
#> 515 514    695 8.832014
#> 516 515    978 8.782027
#> 517 516    825 8.613558
#> 518 517    804 8.683670
#> 519 518   1578 8.518555
#> 520 519   1254 8.709915
#> 521 520   1241 8.902363
#> 522 521   1233 8.971729
#> 523 522    532 9.159905
#> 524 523    764 9.566515
#> 525 524    764 8.567977
#> 526 525    775 8.927364
#> 527 526   1373 8.676095
#> 528 527    856 8.882091
#> 529 528    770 9.183586
#> 530 529   1240 8.960975
#> 531 530    789 9.051231
#> 532 531    413 9.334056
#> 533 532   2629 9.287164
#> 534 533    315 8.441291
#> 535 534    330 8.995834
#> 536 535    330 8.614272
#> 537 536    312 8.193707
#> 538 537    441 9.171864
#> 539 538   3215 9.012588
#> 540 539   1216 9.058574
#> 541 540   1845 8.959426
#> 542 541    560 8.760891
#> 543 542   1017 8.822383
#> 544 543   1517 8.825083
#> 545 544    795 8.573813
#> 546 545    800 8.831166
#> 547 546    272 8.440732
#> 548 547    302 8.403643
#> 549 548    313 8.459872
#> 550 549   2118 8.801461
#> 551 550    307 9.150533
#> 552 551    290 9.084289
#> 553 552   1156 8.886216
#> 554 553   1056 8.733054
#> 555 554   1064 9.332902
#> 556 555   2150 8.967343
#> 557 556    708 8.709742
#> 558 557   1076 8.834391
#> 559 558   2047 8.583260
#> 560 559   1952 8.865736
#> 561 560    791 9.070795
#> 562 561   1620 9.115580
#> 563 562   1636 8.821128
#> 564 563   1297 8.970412
#> 565 564   1273 9.033235
#> 566 565   1262 8.873502
#> 567 566   1302 8.779937
#> 568 567   1289 8.664114
#> 569 568   1258 8.407936
#> 570 569   1285 8.847171
#> 571 570   1259 8.695943
#> 572 571   1499 8.827880
#> 573 572   1475 8.661543
#> 574 573    596 9.568070
#> 575 574    548 8.578695
#> 576 575    532 8.555382
#> 577 576    598 9.435390
#> 578 577    581 8.700295
#> 579 578    526 8.723798
#> 580 579    550 8.672312
#> 581 580    538 9.056936
#> 582 581    574 8.859590
#> 583 582    872 9.373510
#> 584 583    613 8.388770
#> 585 584    335 8.579799
#> 586 585    449 9.220729
#> 587 586    783 8.935229
#> 588 587    647 9.207834
#> 589 588    986 9.220834
#> 590 589   1711 9.457744
#> 591 590   1822 8.837121
#> 592 591    648 8.817019
#> 593 592    809 8.761863
#> 594 593   1346 9.236519
#> 595 594    642 8.889181
#> 596 595    517 8.708051
#> 597 596   1080 8.615039
#> 598 597    628 9.028875
#> 599 598    662 8.727714
#> 600 599   1119 8.914173
#> 601 600   1168 9.188031
#> 602 601    425 8.519601
#> 603 602    385 8.787076
#> 604 603    849 8.668913
#> 605 604   1062 8.828088
#> 606 605   1266 9.125781
#> 607 606   1353 9.354818
#> 608 607    653 9.236310
#> 609 608    625 8.770718
#> 610 609    560 9.122981
#> 611 610   1008 8.644164
#> 612 611   1194 8.630975
#> 613 612   3685 8.995734
#> 614 613   1168 8.793594
#> 615 614   1217 8.275041
#> 616 615   1002 9.138980
#> 617 616   1188 8.655409
#> 618 617    577 8.796592
#> 619 618    555 8.871667
#> 620 619    555 8.543995
#> 621 620    870 8.830500
#> 622 621   1101 8.637669
#> 623 622   1977 9.293509
#> 624 623   2751 9.046116
#> 625 624    381 9.242889