Consume a FASTQ file and collate basic characteristic observations

This method is a fast C++ implementation of a FASTQ parser aimed to perform a quick collation of summary_statistic information that may be used to make decisions as to how FASTQ sequence collections are crafted.

fishy_fastq(fastq)

Value

DataFrame of FASTQ based sequence observations

Examples

fastq <- system.file("extdata", "example.fastq.gz", package = "floundeR")
fishy_fastq(fastq)

#> Reading beyond end of file ...

#>      id sequence_length_template mean_qscore_template passes_filtering
#> 1     0                      481             8.576590             TRUE
#> 2     1                      700             8.938814             TRUE
#> 3     2                      722             8.706461             TRUE
#> 4     3                      402             8.684725             TRUE
#> 5     4                      840             8.765314             TRUE
#> 6     5                      351             8.552162             TRUE
#> 7     6                      780             9.181071             TRUE
#> 8     7                      784             9.070897             TRUE
#> 9     8                      803             8.671617             TRUE
#> 10    9                      769             8.458450             TRUE
#> 11   10                      782             8.561163             TRUE
#> 12   11                      780             9.134631             TRUE
#> 13   12                     1190             9.151952             TRUE
#> 14   13                     1791             8.859815             TRUE
#> 15   14                      725             8.575019             TRUE
#> 16   15                      726             9.166060             TRUE
#> 17   16                      754             9.552390             TRUE
#> 18   17                      780             8.885683             TRUE
#> 19   18                      752             9.174519             TRUE
#> 20   19                      678             8.964424             TRUE
#> 21   20                      546             8.769598             TRUE
#> 22   21                      307             8.288609             TRUE
#> 23   22                     1096             8.798040             TRUE
#> 24   23                      728             8.621111             TRUE
#> 25   24                      601             9.126060             TRUE
#> 26   25                      909             8.610812             TRUE
#> 27   26                     1122             9.252459             TRUE
#> 28   27                     1000             8.649443             TRUE
#> 29   28                      482             9.874456             TRUE
#> 30   29                     3106             9.298542             TRUE
#> 31   30                     4027             8.637724             TRUE
#> 32   31                     1647             8.737156             TRUE
#> 33   32                     1667             9.419045             TRUE
#> 34   33                     1782             8.702462             TRUE
#> 35   34                     1268             8.597244             TRUE
#> 36   35                     1287             8.831912             TRUE
#> 37   36                     1109             9.559415             TRUE
#> 38   37                      516             8.829993             TRUE
#> 39   38                      519             9.965189             TRUE
#> 40   39                      718             9.239699             TRUE
#> 41   40                     1799             8.787300             TRUE
#> 42   41                     1543             8.675742             TRUE
#> 43   42                     1558             9.041159             TRUE
#> 44   43                      970             8.549458             TRUE
#> 45   44                      618             9.518017             TRUE
#> 46   45                      684             8.385486             TRUE
#> 47   46                      872             8.624417             TRUE
#> 48   47                      860             8.708627             TRUE
#> 49   48                      877             9.152344             TRUE
#> 50   49                      761             8.537343             TRUE
#> 51   50                      779             8.797213             TRUE
#> 52   51                      624             8.766164             TRUE
#> 53   52                      640             8.781042             TRUE
#> 54   53                      499             8.585148             TRUE
#> 55   54                     1634             8.900698             TRUE
#> 56   55                     2053             9.110444             TRUE
#> 57   56                      561             9.033309             TRUE
#> 58   57                      528             9.710940             TRUE
#> 59   58                      544             9.138390             TRUE
#> 60   59                      495             9.230347             TRUE
#> 61   60                      555             9.087939             TRUE
#> 62   61                      545             8.667213             TRUE
#> 63   62                      352             9.168268             TRUE
#> 64   63                      534             8.899837             TRUE
#> 65   64                     1022             8.583999             TRUE
#> 66   65                     1452             8.914466             TRUE
#> 67   66                      705             8.781941             TRUE
#> 68   67                      629             8.561754             TRUE
#> 69   68                      566             8.791556             TRUE
#> 70   69                      627             8.992131             TRUE
#> 71   70                      630             8.966647             TRUE
#> 72   71                      564             8.939142             TRUE
#> 73   72                      555             8.902081             TRUE
#> 74   73                      558             8.587989             TRUE
#> 75   74                      895             8.379152             TRUE
#> 76   75                      528             9.551015             TRUE
#> 77   76                      583             9.451621             TRUE
#> 78   77                      611             9.131329             TRUE
#> 79   78                      828             8.572431             TRUE
#> 80   79                     1766             9.430747             TRUE
#> 81   80                     1237             9.289551             TRUE
#> 82   81                      693             8.811421             TRUE
#> 83   82                      382             8.894040             TRUE
#> 84   83                      963             8.754269             TRUE
#> 85   84                      805             8.823750             TRUE
#> 86   85                     1163             8.951216             TRUE
#> 87   86                      970             8.843556             TRUE
#> 88   87                      925             9.168600             TRUE
#> 89   88                      949             8.690817             TRUE
#> 90   89                      932             8.889891             TRUE
#> 91   90                      940             8.832452             TRUE
#> 92   91                      901             8.923236             TRUE
#> 93   92                      936             8.859321             TRUE
#> 94   93                      943             8.870780             TRUE
#> 95   94                      980             9.188061             TRUE
#> 96   95                      907             8.805997             TRUE
#> 97   96                      878             9.357055             TRUE
#> 98   97                      739             8.722930             TRUE
#> 99   98                     1254             8.733398             TRUE
#> 100  99                      662             8.753577             TRUE
#> 101 100                     2203             9.294639             TRUE
#> 102 101                      664             9.009345             TRUE
#> 103 102                      675             9.159626             TRUE
#> 104 103                     1213             8.866251             TRUE
#> 105 104                     1236             9.026835             TRUE
#> 106 105                     1068             9.263474             TRUE
#> 107 106                     1762             9.445290             TRUE
#> 108 107                     1569             9.044123             TRUE
#> 109 108                     1021             8.530767             TRUE
#> 110 109                      987             9.531640             TRUE
#> 111 110                     3278             9.188058             TRUE
#> 112 111                      758             8.689724             TRUE
#> 113 112                     1758             8.795593             TRUE
#> 114 113                      916             8.978864             TRUE
#> 115 114                     1015             9.262681             TRUE
#> 116 115                     2092             9.271888             TRUE
#> 117 116                     2024             9.273611             TRUE
#> 118 117                     2034             9.334114             TRUE
#> 119 118                      799             8.514698             TRUE
#> 120 119                      961             8.807415             TRUE
#> 121 120                      532             8.713382             TRUE
#> 122 121                     2803             9.111441             TRUE
#> 123 122                      691             9.153989             TRUE
#> 124 123                     1809             8.610636             TRUE
#> 125 124                     1431             9.243698             TRUE
#> 126 125                     1730             8.672883             TRUE
#> 127 126                     1829             8.631370             TRUE
#> 128 127                     1740             8.878706             TRUE
#> 129 128                      333             8.660597             TRUE
#> 130 129                      618             8.809836             TRUE
#> 131 130                      874             8.598460             TRUE
#> 132 131                     1149             8.616637             TRUE
#> 133 132                      881             9.027526             TRUE
#> 134 133                     1792             9.154873             TRUE
#> 135 134                      538             8.850554             TRUE
#> 136 135                      459             8.985008             TRUE
#> 137 136                      588             8.674768             TRUE
#> 138 137                      831             8.673054             TRUE
#> 139 138                      563             9.239863             TRUE
#> 140 139                      606             9.284954             TRUE
#> 141 140                     1330             8.675208             TRUE
#> 142 141                      475             9.601694             TRUE
#> 143 142                      651             9.318546             TRUE
#> 144 143                      954             8.614737             TRUE
#> 145 144                      562             8.593277             TRUE
#> 146 145                     1906             8.952943             TRUE
#> 147 146                      616             8.636640             TRUE
#> 148 147                     1473             8.839078             TRUE
#> 149 148                      868             8.439160             TRUE
#> 150 149                     1296             8.962461             TRUE
#> 151 150                      586             8.880845             TRUE
#> 152 151                     3099             9.046738             TRUE
#> 153 152                     1174             8.915950             TRUE
#> 154 153                     1040             8.896862             TRUE
#> 155 154                     3058             8.763107             TRUE
#> 156 155                      956             8.978460             TRUE
#> 157 156                     2257             8.725055             TRUE
#> 158 157                     2713             9.066904             TRUE
#> 159 158                     2922             8.399341             TRUE
#> 160 159                     2612             8.677419             TRUE
#> 161 160                     2193             8.959013             TRUE
#> 162 161                     2705             9.426265             TRUE
#> 163 162                      748             8.674518             TRUE
#> 164 163                      699             9.220833             TRUE
#> 165 164                     1150             8.679215             TRUE
#> 166 165                      928             8.744246             TRUE
#> 167 166                      931             8.945282             TRUE
#> 168 167                      899             9.121024             TRUE
#> 169 168                     1002             8.918637             TRUE
#> 170 169                     2678             8.913277             TRUE
#> 171 170                      912             8.397955             TRUE
#> 172 171                     1511             8.863208             TRUE
#> 173 172                     2072             8.806537             TRUE
#> 174 173                     2197             8.748316             TRUE
#> 175 174                     1411             9.369255             TRUE
#> 176 175                     1231             8.652401             TRUE
#> 177 176                     1101             8.869453             TRUE
#> 178 177                     1681             9.044779             TRUE
#> 179 178                     1134             9.895219             TRUE
#> 180 179                     1726             8.659587             TRUE
#> 181 180                     1670             8.924303             TRUE
#> 182 181                     1338             8.716558             TRUE
#> 183 182                     1692             8.922948             TRUE
#> 184 183                     1507             8.880793             TRUE
#> 185 184                     1659             8.969820             TRUE
#> 186 185                     1740             8.953312             TRUE
#> 187 186                      432             8.816593             TRUE
#> 188 187                     1741             8.799508             TRUE
#> 189 188                     1704             8.680899             TRUE
#> 190 189                     1476             8.719795             TRUE
#> 191 190                     1734             8.932984             TRUE
#> 192 191                     1676             9.147722             TRUE
#> 193 192                     1090             8.956491             TRUE
#> 194 193                     1083             9.116109             TRUE
#> 195 194                     1237             9.169847             TRUE
#> 196 195                      758             8.603536             TRUE
#> 197 196                      702             9.061111             TRUE
#> 198 197                      700             8.614972             TRUE
#> 199 198                      592             8.770947             TRUE
#> 200 199                      614             8.582882             TRUE
#> 201 200                     1108             8.922046             TRUE
#> 202 201                     3317             8.699775             TRUE
#> 203 202                      767             8.973541             TRUE
#> 204 203                     2635             8.903270             TRUE
#> 205 204                     1895             8.877907             TRUE
#> 206 205                     1945             8.670979             TRUE
#> 207 206                      586             8.854418             TRUE
#> 208 207                     1175             8.860614             TRUE
#> 209 208                     1459             8.832618             TRUE
#> 210 209                     1449             8.921386             TRUE
#> 211 210                     2060             9.016544             TRUE
#> 212 211                     2034             9.465502             TRUE
#> 213 212                     1407             9.173653             TRUE
#> 214 213                     1545             8.561141             TRUE
#> 215 214                     1140             8.897316             TRUE
#> 216 215                      805             8.919828             TRUE
#> 217 216                     1105             8.957165             TRUE
#> 218 217                     1084             9.147075             TRUE
#> 219 218                      734             8.442271             TRUE
#> 220 219                      704             8.649677             TRUE
#> 221 220                      496             8.775745             TRUE
#> 222 221                      488             8.810199             TRUE
#> 223 222                      360             8.765608             TRUE
#> 224 223                      851             9.094712             TRUE
#> 225 224                      993             8.676013             TRUE
#> 226 225                     1855             9.159714             TRUE
#> 227 226                      514             8.590481             TRUE
#> 228 227                     1521             8.923656             TRUE
#> 229 228                      562             8.529856             TRUE
#> 230 229                     1329             9.104410             TRUE
#> 231 230                      601             8.943021             TRUE
#> 232 231                     1073             8.839932             TRUE
#> 233 232                      534             8.835434             TRUE
#> 234 233                      365             8.512503             TRUE
#> 235 234                     2791             9.131243             TRUE
#> 236 235                      512             8.391978             TRUE
#> 237 236                     1835             8.731873             TRUE
#> 238 237                     1010             9.240406             TRUE
#> 239 238                     1089             8.544054             TRUE
#> 240 239                     1062             9.016390             TRUE
#> 241 240                     1060             8.952624             TRUE
#> 242 241                     1113             8.686292             TRUE
#> 243 242                     1091             8.860182             TRUE
#> 244 243                     1079             8.880504             TRUE
#> 245 244                     1050             8.844290             TRUE
#> 246 245                     1056             9.336447             TRUE
#> 247 246                     1033             8.847164             TRUE
#> 248 247                      483             8.933823             TRUE
#> 249 248                     1796             8.566093             TRUE
#> 250 249                     1816             9.040464             TRUE
#> 251 250                      911             8.817411             TRUE
#> 252 251                      651             8.937114             TRUE
#> 253 252                     1295             8.959100             TRUE
#> 254 253                     1521             8.749072             TRUE
#> 255 254                     1557             8.902612             TRUE
#> 256 255                     1118             9.066720             TRUE
#> 257 256                     1218             8.654493             TRUE
#> 258 257                     1210             9.031938             TRUE
#> 259 258                      923             8.522453             TRUE
#> 260 259                     1059             8.932382             TRUE
#> 261 260                      944             8.564706             TRUE
#> 262 261                     1597             9.122895             TRUE
#> 263 262                     1357             8.970204             TRUE
#> 264 263                     1289             9.096166             TRUE
#> 265 264                     1303             9.231134             TRUE
#> 266 265                      846             9.000420             TRUE
#> 267 266                     1044             8.505878             TRUE
#> 268 267                     1258             8.580690             TRUE
#> 269 268                      563             9.114577             TRUE
#> 270 269                      844             8.664027             TRUE
#> 271 270                     1606             8.613220             TRUE
#> 272 271                     1037             8.606793             TRUE
#> 273 272                     1378             8.740752             TRUE
#> 274 273                     1923             9.183908             TRUE
#> 275 274                     1551             8.768124             TRUE
#> 276 275                     1069             8.785294             TRUE
#> 277 276                      965             9.157911             TRUE
#> 278 277                      516             8.599988             TRUE
#> 279 278                     1354             8.575545             TRUE
#> 280 279                      612             8.594377             TRUE
#> 281 280                      342             9.194752             TRUE
#> 282 281                     1035             9.017363             TRUE
#> 283 282                     1154             9.038597             TRUE
#> 284 283                     1365             9.038660             TRUE
#> 285 284                     2311             8.839053             TRUE
#> 286 285                      858             9.229750             TRUE
#> 287 286                      824             8.872558             TRUE
#> 288 287                      561             9.472907             TRUE
#> 289 288                     2034             8.899984             TRUE
#> 290 289                      389             9.092170             TRUE
#> 291 290                     1305             9.010448             TRUE
#> 292 291                     1237             9.154613             TRUE
#> 293 292                      531             8.825195             TRUE
#> 294 293                      486             8.945825             TRUE
#> 295 294                     2824             8.676455             TRUE
#> 296 295                     1276             8.980473             TRUE
#> 297 296                     1258             8.995400             TRUE
#> 298 297                      966             8.963307             TRUE
#> 299 298                      871             8.959463             TRUE
#> 300 299                      855             9.058735             TRUE
#> 301 300                      871             8.963910             TRUE
#> 302 301                      836             9.275776             TRUE
#> 303 302                     1284             8.836848             TRUE
#> 304 303                      896             8.501498             TRUE
#> 305 304                     1654             8.742943             TRUE
#> 306 305                      729             9.043782             TRUE
#> 307 306                      335             8.674843             TRUE
#> 308 307                      922             8.932634             TRUE
#> 309 308                      823             8.962562             TRUE
#> 310 309                      753             9.267746             TRUE
#> 311 310                      893             8.687301             TRUE
#> 312 311                     1120             8.717583             TRUE
#> 313 312                     1151             9.056684             TRUE
#> 314 313                      640             8.987849             TRUE
#> 315 314                     1105             8.762798             TRUE
#> 316 315                     1142             9.610859             TRUE
#> 317 316                     1137             8.801708             TRUE
#> 318 317                     1163             8.686275             TRUE
#> 319 318                     1150             8.623515             TRUE
#> 320 319                     1164             8.874183             TRUE
#> 321 320                     1141             8.780610             TRUE
#> 322 321                     1148             9.015638             TRUE
#> 323 322                     1139             8.771626             TRUE
#> 324 323                     1139             8.563053             TRUE
#> 325 324                     1131             8.971964             TRUE
#> 326 325                      943             8.925531             TRUE
#> 327 326                     1151             9.034090             TRUE
#> 328 327                     1109             9.156023             TRUE
#> 329 328                     1129             9.027990             TRUE
#> 330 329                     1151             8.947779             TRUE
#> 331 330                     1130             8.598717             TRUE
#> 332 331                     1124             9.160414             TRUE
#> 333 332                     1128             8.984138             TRUE
#> 334 333                     1137             8.711840             TRUE
#> 335 334                      921             8.942837             TRUE
#> 336 335                      633             9.006188             TRUE
#> 337 336                     1115             8.478238             TRUE
#> 338 337                     1144             8.593292             TRUE
#> 339 338                     1132             9.032812             TRUE
#> 340 339                     1132             8.749082             TRUE
#> 341 340                     1115             8.912217             TRUE
#> 342 341                     1128             8.885962             TRUE
#> 343 342                     1147             9.084100             TRUE
#> 344 343                     1134             9.276955             TRUE
#> 345 344                     1102             8.630629             TRUE
#> 346 345                     1136             9.562294             TRUE
#> 347 346                     1116             8.775780             TRUE
#> 348 347                     1090             8.802011             TRUE
#> 349 348                     1117             8.728036             TRUE
#> 350 349                      615             8.866208             TRUE
#> 351 350                     1936             8.918410             TRUE
#> 352 351                      477             9.306606             TRUE
#> 353 352                      479             8.585401             TRUE
#> 354 353                      462             8.618173             TRUE
#> 355 354                     2538             8.770000             TRUE
#> 356 355                     5946             9.232093             TRUE
#> 357 356                     2992             9.061294             TRUE
#> 358 357                     2453             9.606964             TRUE
#> 359 358                     3141             8.764297             TRUE
#> 360 359                     3159             9.013587             TRUE
#> 361 360                     1967             8.972908             TRUE
#> 362 361                     2136             8.633924             TRUE
#> 363 362                      378             8.904277             TRUE
#> 364 363                     1036             9.395297             TRUE
#> 365 364                      949             8.960071             TRUE
#> 366 365                     1006             8.844896             TRUE
#> 367 366                     1008             8.711229             TRUE
#> 368 367                      985             8.919939             TRUE
#> 369 368                      978             8.566126             TRUE
#> 370 369                      975             8.844392             TRUE
#> 371 370                      969             9.210991             TRUE
#> 372 371                     1003             8.756380             TRUE
#> 373 372                      863             8.792306             TRUE
#> 374 373                      684             8.981501             TRUE
#> 375 374                      677             9.091798             TRUE
#> 376 375                      658             8.735104             TRUE
#> 377 376                      984             8.802644             TRUE
#> 378 377                      672             8.848940             TRUE
#> 379 378                      681             8.858216             TRUE
#> 380 379                      653             9.400763             TRUE
#> 381 380                      659             8.895428             TRUE
#> 382 381                      652             8.926518             TRUE
#> 383 382                      657             8.762445             TRUE
#> 384 383                      674             9.052426             TRUE
#> 385 384                      669             8.722652             TRUE
#> 386 385                      675             8.743557             TRUE
#> 387 386                      660             9.074414             TRUE
#> 388 387                      642             8.493161             TRUE
#> 389 388                      631             8.702148             TRUE
#> 390 389                      691             9.084943             TRUE
#> 391 390                      663             8.860919             TRUE
#> 392 391                      679             9.067593             TRUE
#> 393 392                      652             8.978120             TRUE
#> 394 393                      656             8.951246             TRUE
#> 395 394                      660             8.982062             TRUE
#> 396 395                      682             9.247323             TRUE
#> 397 396                      681             8.907953             TRUE
#> 398 397                      697             8.854692             TRUE
#> 399 398                      655             9.153746             TRUE
#> 400 399                      609             8.863561             TRUE
#> 401 400                      639             9.220086             TRUE
#> 402 401                      659             8.532218             TRUE
#> 403 402                      669             8.954770             TRUE
#> 404 403                      663             8.512855             TRUE
#> 405 404                      646             9.342391             TRUE
#> 406 405                      691             8.889667             TRUE
#> 407 406                      662             9.125675             TRUE
#> 408 407                      687             9.007006             TRUE
#> 409 408                      690             9.171832             TRUE
#> 410 409                      583             8.563015             TRUE
#> 411 410                      534             9.376023             TRUE
#> 412 411                     2681             8.866687             TRUE
#> 413 412                      429             8.796249             TRUE
#> 414 413                     1115             8.836318             TRUE
#> 415 414                     1104             8.899658             TRUE
#> 416 415                     1379             8.780057             TRUE
#> 417 416                     1752             8.710213             TRUE
#> 418 417                     1090             9.215020             TRUE
#> 419 418                     1564             8.685073             TRUE
#> 420 419                     1578             8.965851             TRUE
#> 421 420                     1280             8.915881             TRUE
#> 422 421                     2345             8.805212             TRUE
#> 423 422                     1571             8.711089             TRUE
#> 424 423                     1514             8.860684             TRUE
#> 425 424                     1525             8.793452             TRUE
#> 426 425                     1333             9.269311             TRUE
#> 427 426                     2111             9.171345             TRUE
#> 428 427                     1900             8.915723             TRUE
#> 429 428                      767             8.379281             TRUE
#> 430 429                     1741             8.744231             TRUE
#> 431 430                     1277             8.916176             TRUE
#> 432 431                     1010             8.936346             TRUE
#> 433 432                      943             8.698577             TRUE
#> 434 433                     1904             9.212870             TRUE
#> 435 434                     2399             8.808081             TRUE
#> 436 435                     2392             9.006919             TRUE
#> 437 436                      549             8.857136             TRUE
#> 438 437                      499             9.226223             TRUE
#> 439 438                      520             8.891541             TRUE
#> 440 439                      637             9.488916             TRUE
#> 441 440                      526             9.098991             TRUE
#> 442 441                      476             9.642491             TRUE
#> 443 442                      553             8.893782             TRUE
#> 444 443                      539             9.403773             TRUE
#> 445 444                      571             8.783266             TRUE
#> 446 445                     1317             8.967943             TRUE
#> 447 446                      530             9.083784             TRUE
#> 448 447                      550             9.016772             TRUE
#> 449 448                      549             9.133563             TRUE
#> 450 449                      695             9.583085             TRUE
#> 451 450                      353             8.423545             TRUE
#> 452 451                      948             9.225145             TRUE
#> 453 452                      906             8.729439             TRUE
#> 454 453                      523             9.521440             TRUE
#> 455 454                      563             8.596677             TRUE
#> 456 455                      535             9.052860             TRUE
#> 457 456                     1320             8.675113             TRUE
#> 458 457                      575             8.209250             TRUE
#> 459 458                      562             8.595451             TRUE
#> 460 459                      585             8.721528             TRUE
#> 461 460                      566             9.056266             TRUE
#> 462 461                      565             8.462854             TRUE
#> 463 462                      573             8.319966             TRUE
#> 464 463                      564             8.811651             TRUE
#> 465 464                      577             8.811384             TRUE
#> 466 465                      587             8.763002             TRUE
#> 467 466                      455             8.592346             TRUE
#> 468 467                      860             9.394502             TRUE
#> 469 468                     1382             8.929006             TRUE
#> 470 469                     1358             9.421250             TRUE
#> 471 470                      945             9.277841             TRUE
#> 472 471                      838             8.868300             TRUE
#> 473 472                     1361             8.596629             TRUE
#> 474 473                     1995             8.956285             TRUE
#> 475 474                      957             8.867213             TRUE
#> 476 475                      978             8.962595             TRUE
#> 477 476                      954             9.152226             TRUE
#> 478 477                      719             8.975530             TRUE
#> 479 478                      599             9.127474             TRUE
#> 480 479                      357             8.456306             TRUE
#> 481 480                      650             8.980589             TRUE
#> 482 481                      643             9.426341             TRUE
#> 483 482                      571             8.898402             TRUE
#> 484 483                      605             8.593619             TRUE
#> 485 484                      600             9.177834             TRUE
#> 486 485                      570             9.355305             TRUE
#> 487 486                      546             8.686034             TRUE
#> 488 487                      589             8.809509             TRUE
#> 489 488                      607             8.980917             TRUE
#> 490 489                      465             9.479959             TRUE
#> 491 490                      605             8.616143             TRUE
#> 492 491                      604             8.699931             TRUE
#> 493 492                      592             8.440726             TRUE
#> 494 493                      566             8.789350             TRUE
#> 495 494                     1034             9.169764             TRUE
#> 496 495                     1057             8.908205             TRUE
#> 497 496                     1041             8.879029             TRUE
#> 498 497                     1033             9.698854             TRUE
#> 499 498                     1036             9.131104             TRUE
#> 500 499                     1044             9.776451             TRUE
#> 501 500                     1089             8.809536             TRUE
#> 502 501                     1025             8.708831             TRUE
#> 503 502                      603             8.300916             TRUE
#> 504 503                     1018             8.804236             TRUE
#> 505 504                     1030             9.078977             TRUE
#> 506 505                     1050             8.625113             TRUE
#> 507 506                     1032             8.644139             TRUE
#> 508 507                     1063             9.027771             TRUE
#> 509 508                      173             8.720148             TRUE
#> 510 509                      623             8.915773             TRUE
#> 511 510                      699             8.463956             TRUE
#> 512 511                     1060             9.382037             TRUE
#> 513 512                     1006             9.081835             TRUE
#> 514 513                     1043             9.125876             TRUE
#> 515 514                      695             8.832014             TRUE
#> 516 515                      978             8.782027             TRUE
#> 517 516                      825             8.613558             TRUE
#> 518 517                      804             8.683670             TRUE
#> 519 518                     1578             8.518555             TRUE
#> 520 519                     1254             8.709915             TRUE
#> 521 520                     1241             8.902363             TRUE
#> 522 521                     1233             8.971729             TRUE
#> 523 522                      532             9.159905             TRUE
#> 524 523                      764             9.566515             TRUE
#> 525 524                      764             8.567977             TRUE
#> 526 525                      775             8.927364             TRUE
#> 527 526                     1373             8.676095             TRUE
#> 528 527                      856             8.882091             TRUE
#> 529 528                      770             9.183586             TRUE
#> 530 529                     1240             8.960975             TRUE
#> 531 530                      789             9.051231             TRUE
#> 532 531                      413             9.334056             TRUE
#> 533 532                     2629             9.287164             TRUE
#> 534 533                      315             8.441291             TRUE
#> 535 534                      330             8.995834             TRUE
#> 536 535                      330             8.614272             TRUE
#> 537 536                      312             8.193707             TRUE
#> 538 537                      441             9.171864             TRUE
#> 539 538                     3215             9.012588             TRUE
#> 540 539                     1216             9.058574             TRUE
#> 541 540                     1845             8.959426             TRUE
#> 542 541                      560             8.760891             TRUE
#> 543 542                     1017             8.822383             TRUE
#> 544 543                     1517             8.825083             TRUE
#> 545 544                      795             8.573813             TRUE
#> 546 545                      800             8.831166             TRUE
#> 547 546                      272             8.440732             TRUE
#> 548 547                      302             8.403643             TRUE
#> 549 548                      313             8.459872             TRUE
#> 550 549                     2118             8.801461             TRUE
#> 551 550                      307             9.150533             TRUE
#> 552 551                      290             9.084289             TRUE
#> 553 552                     1156             8.886216             TRUE
#> 554 553                     1056             8.733054             TRUE
#> 555 554                     1064             9.332902             TRUE
#> 556 555                     2150             8.967343             TRUE
#> 557 556                      708             8.709742             TRUE
#> 558 557                     1076             8.834391             TRUE
#> 559 558                     2047             8.583260             TRUE
#> 560 559                     1952             8.865736             TRUE
#> 561 560                      791             9.070795             TRUE
#> 562 561                     1620             9.115580             TRUE
#> 563 562                     1636             8.821128             TRUE
#> 564 563                     1297             8.970412             TRUE
#> 565 564                     1273             9.033235             TRUE
#> 566 565                     1262             8.873502             TRUE
#> 567 566                     1302             8.779937             TRUE
#> 568 567                     1289             8.664114             TRUE
#> 569 568                     1258             8.407936             TRUE
#> 570 569                     1285             8.847171             TRUE
#> 571 570                     1259             8.695943             TRUE
#> 572 571                     1499             8.827880             TRUE
#> 573 572                     1475             8.661543             TRUE
#> 574 573                      596             9.568070             TRUE
#> 575 574                      548             8.578695             TRUE
#> 576 575                      532             8.555382             TRUE
#> 577 576                      598             9.435390             TRUE
#> 578 577                      581             8.700295             TRUE
#> 579 578                      526             8.723798             TRUE
#> 580 579                      550             8.672312             TRUE
#> 581 580                      538             9.056936             TRUE
#> 582 581                      574             8.859590             TRUE
#> 583 582                      872             9.373510             TRUE
#> 584 583                      613             8.388770             TRUE
#> 585 584                      335             8.579799             TRUE
#> 586 585                      449             9.220729             TRUE
#> 587 586                      783             8.935229             TRUE
#> 588 587                      647             9.207834             TRUE
#> 589 588                      986             9.220834             TRUE
#> 590 589                     1711             9.457744             TRUE
#> 591 590                     1822             8.837121             TRUE
#> 592 591                      648             8.817019             TRUE
#> 593 592                      809             8.761863             TRUE
#> 594 593                     1346             9.236519             TRUE
#> 595 594                      642             8.889181             TRUE
#> 596 595                      517             8.708051             TRUE
#> 597 596                     1080             8.615039             TRUE
#> 598 597                      628             9.028875             TRUE
#> 599 598                      662             8.727714             TRUE
#> 600 599                     1119             8.914173             TRUE
#> 601 600                     1168             9.188031             TRUE
#> 602 601                      425             8.519601             TRUE
#> 603 602                      385             8.787076             TRUE
#> 604 603                      849             8.668913             TRUE
#> 605 604                     1062             8.828088             TRUE
#> 606 605                     1266             9.125781             TRUE
#> 607 606                     1353             9.354818             TRUE
#> 608 607                      653             9.236310             TRUE
#> 609 608                      625             8.770718             TRUE
#> 610 609                      560             9.122981             TRUE
#> 611 610                     1008             8.644164             TRUE
#> 612 611                     1194             8.630975             TRUE
#> 613 612                     3685             8.995734             TRUE
#> 614 613                     1168             8.793594             TRUE
#> 615 614                     1217             8.275041             TRUE
#> 616 615                     1002             9.138980             TRUE
#> 617 616                     1188             8.655409             TRUE
#> 618 617                      577             8.796592             TRUE
#> 619 618                      555             8.871667             TRUE
#> 620 619                      555             8.543995             TRUE
#> 621 620                      870             8.830500             TRUE
#> 622 621                     1101             8.637669             TRUE
#> 623 622                     1977             9.293509             TRUE
#> 624 623                     2751             9.046116             TRUE
#> 625 624                      381             9.242889             TRUE